feat(signup, auth): validate form inputs
Implements both server-side and client-side form validation for the
signup form. On the client side, the form is validated in real-time as
the user enters keyboard inputs to the fields. Once the "Sign Up" button
is clicked, the form is sent to and validated on the server. Server-side
validation results are returned and displayed at the top of the form.

Uses Zod as the schema validation library.