feat(modalWrapper): Created a modal wrapper component. Since this "modal" is being used multiple times in different areas, creating a wrapper would allow for modularity instead of recreating the "modal" multiple times.
refactor(templateModal): Used the modal wrapper.
Used the modal wrapper to wrap create and view sections for the modal.

The events tab will still yet to be done. When user clicks and adds an event, the events is added locally not saved into database yet. Another thing is that the added event will have to be clickable. When clicked it should open a modal which is editable maybe can include autosave as well.