test: improve performance of tests
Previously, we had to generate a new database and migrate it to the
schemas for every single test, which is slow. This commit makes it so
that a new database is generated only for each test set (i.e., a single
file) - this database is shared by each test in the set. After each test,
the `reset()` function from `drizzle-seed` is used to reset the database
to an empty state.